requirements.txt
Django==1.10.5
djangorestframework==3.5.3
python-telegram-bot==5.3.0
redis==2.10.5
